Removed NSLogs

+++ NSString+Ruby/NSString+Ruby.h
@@ -35,7 +35,7 @@
 -(BOOL)includes:(NSString*)include;
 -(NSInteger)index:(NSString*)pattern;
 -(NSInteger)index:(NSString*)pattern offset:(int)offset;
--(NSString*)insert:(long)index string:(NSString*)string;
+-(NSString*)insert:(int)index string:(NSString*)string;
 -(NSString*)inspect;
 -(BOOL)isASCII;
 -(BOOL)isEmpty;
